Make modes customizable and editable
As per request of some users, sometimes new modes or missing modes had to be
added into the code.

As that hardcoded lists are not eary maintainable, we created a new table with
all modes and submodes listed and make this list editable.

Now one can activate or deativate modes from beeing shown within the select-
list in QSO Window.