Add script to automate the update process more
We recommend creating an update script in the install notes so why not provide one that works for the environment we refer to there (Debian/Ubuntu)? This script automates the git pull and setting of permissions on writable directories. It has a bit of error handling for safety and allows the admin to choose whether to update to the tip of master (current behaviour) or to the latest tag.