Add the gridsquare from LotW QSL to the logbook
The gridsquare from the QSL will now update the log if the gridsquare
was not previously populated, or if the LotW gridsquare is more
specific than what previously existed.  Previously the gridsquare was
not being used from the LotW QSL record.

This will enable the map above the logbook to more accurately display
where the QSL was with instead of default values.

 - Add 'qsl_gridsquare' to the logbook_model.php:lotw_update
   implementation.
 - Refactor code in lotw_update to remove duplicate code.
 - Add gridsquare into the LotW output table on what is imported.