Correct user date formating on QSO pages (Live/Post)
The date field will now be populated from the user's preferences, if
this is not set it will fallback to the file config setting.

-Move date format detection to earlier in the index.php file to allow a
single lookup of the format for the entire file.
-Convert hardcode date string to format selection in index.php to allow
the custom format to show on the Post QSO page.
-Convert php format string to jclock format string in footer.php to
allow the custom format to show on the Live QSO page.