Update Distances_model : change "round" to "ceil"
I propose change "round" to "ceil" (Round fractions up).

Because I work on an export plugin in EDI file (IARU REGTEST1), the rounding is systematically higher:
"For the amateur bands up to 10 GHz inclusive, points will be scored on the basis of one point per
kilometer, i.e. the calculated distance in kilometers will be truncated to an integer value and 1 km
will be added. " (excerpt from VHF_Handbook_V9.01)